+++ packages/coding-standards/stylelint/index.cjs
@@ -1,5 +1,0 @@
-module.exports = {
-	classic: require("./classic.cjs"),
-	themeJson: require("./theme-json.cjs"),
-	tailwind: require("./tailwind.cjs")
-};