+++ packages/coding-standards/stylelint/index.mjs
@@ -1,4 +1,0 @@
-import { default as classic } from "./classic.mjs";
-import { default as tailwind } from "./tailwind.mjs";
-import { default as themeJson } from "./theme-json.mjs";
-export default { classic, themeJson, tailwind };