+++ prettier.config.mjs
@@ -1,2 +1,4 @@
-import { prettierConfig } from "@atomicsmash/coding-standards/index.mjs";
+// Don't copy this for projects, this is specifically dogfooding itself.
+// For projects use import { prettierConfig } from "@atomicsmash/coding-standards";
+import prettierConfig from "./packages/coding-standards/dist/prettier/index.js";
 export default prettierConfig;